pyproject.toml
[build-system]
requires = ["setuptools>=45"]
build-backend = "setuptools.build_meta"

[project]
name = "cheshirecat-admin"
version = "0.2.0"
description = "Python package for the Cheshire Cat Admin UI"
authors = [
    {name = "Matteo Cacciola", email = "matteo.cacciola@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cheshirecat-python-sdk",
    "python-dotenv",
    "python-slugify",
    "requests",
    "streamlit",
]

[project.urls]
Homepage = "https://github.com/matteocacciola/cheshirecat-admin"
Repository = "https://github.com/matteocacciola/cheshirecat-admin"
Issues = "https://github.com/matteocacciola/cheshirecat-admin/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
